Tests 	Working on download recovery testing. Propagated previous parameter change to syncServerDownloadsComplete delegate callback to the testing code. Got an additional download/crash recovery test working.
iOS client
	Fixed bug in download recovery. Needed the server file index persisted in order for the .Download mode recovery to operate. In order to persist the server file index, implemented NSCoding for SMServerFile class.

+++ iOSTests/Code/TestBasics.swift
@@ -123,7 +123,7 @@ public class TestBasics {
         SMServerAPI.session.getFileIndex() { (fileIndex, apiResult) in
             if apiResult.error == nil {
                 let result = fileIndex!.filter({
-                    $0.uuid.UUIDString == uuid && $0.sizeBytes == Int32(size)
+                    $0.uuid.UUIDString == uuid && $0.sizeBytes == size
                 })
                 if result.count == 1 {
                     finish()
